bank-analysis-ai-chatbot/backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pc-analyst"
version = "0.1.0"
description = "Private Credit Analyst Tool - backend"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "psycopg[binary]>=3.1",
  "pgvector>=0.2.5",
  "sqlalchemy>=2.0",
  "beautifulsoup4>=4.12",
  "lxml>=5.1",
  "httpx>=0.27",
  "tenacity>=8.2",
  "rank-bm25>=0.2.2",
  "numpy>=1.26",
  "pyyaml>=6.0",
  "anthropic>=0.25",
  "sentence-transformers>=2.6",
  "mcp>=1.0",
  "rich>=13.7",
  "typer>=0.12",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "ruff>=0.3",
]

[project.scripts]
pc-analyst = "pc_analyst.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
pc_analyst = ["taxonomy/*.yaml"]

[tool.ruff]
line-length = 100
target-version = "py311"
